Make C# external video track source a proper source
Change ExternalVideoTrackSource to derive from the same base class as
the device-based video track source, which is renamed to
DeviceVideoTrackSource for clarity. The base class is VideoTrackSource.
